track media steps, move media processing into token processing
