+++ service/persist/token_processing_jobs.go
@@ -41,6 +41,17 @@ type TokenProperties struct {
 	HasDescription  bool `json:"has_description"`
 }
 
+func (t TokenProperties) Value() (driver.Value, error) {
+	return json.Marshal(t)
+}
+
+func (t *TokenProperties) Scan(value interface{}) error {
+	if value == nil {
+		return nil
+	}
+	return json.Unmarshal(value.([]byte), t)
+}
+
 type PipelineStepStatus string
 
 const (
@@ -77,26 +88,44 @@ func (p PipelineStepStatus) MarshalJSON() ([]byte, error) {
 }
 
 type PipelineMetadata struct {
-	MetadataRetrieval            PipelineStepStatus `json:"metadata_retrieval,omitempty"`
-	TokenInfoRetrieval           PipelineStepStatus `json:"token_info_retrieval,omitempty"`
-	MediaURLsRetrieval           PipelineStepStatus `json:"media_urls_retrieval,omitempty"`
-	ContentHeaderValueRetrieval  PipelineStepStatus `json:"content_header_value_retrieval,omitempty"`
-	ReaderRetrieval              PipelineStepStatus `json:"reader_retrieval,omitempty"`
-	OpenseaFallback              PipelineStepStatus `json:"opensea_fallback,omitempty"`
-	DetermineMediaTypeWithReader PipelineStepStatus `json:"determine_media_type_with_reader,omitempty"`
-	AnimationGzip                PipelineStepStatus `json:"animation_gzip,omitempty"`
-	StoreGCP                     PipelineStepStatus `json:"store_gcp,omitempty"`
-	ThumbnailGCP                 PipelineStepStatus `json:"thumbnail_gcp,omitempty"`
-	LiveRenderGCP                PipelineStepStatus `json:"live_render_gcp,omitempty"`
-	RawMediaDetermination        PipelineStepStatus `json:"raw_media_determination,omitempty"`
-	NothingCachedWithErrors      PipelineStepStatus `json:"nothing_cached_errors,omitempty"`
-	NothingCachedWithoutErrors   PipelineStepStatus `json:"nothing_cached_no_errors,omitempty"`
-	CreateMediaFromCachedObjects PipelineStepStatus `json:"create_media_from_cached_objects,omitempty"`
-	CreateRawMedia               PipelineStepStatus `json:"create_raw_media,omitempty"`
-	SetUnknownMediaType          PipelineStepStatus `json:"set_default_media_type,omitempty"`
-	MediaResultComparison        PipelineStepStatus `json:"media_result_comparison,omitempty"`
-	UpdateTokenMetadataDB        PipelineStepStatus `json:"update_token_metadata_db,omitempty"`
-	UpdateJobDB                  PipelineStepStatus `json:"update_job_db,omitempty"`
+	MetadataRetrieval                     PipelineStepStatus `json:"metadata_retrieval,omitempty"`
+	TokenInfoRetrieval                    PipelineStepStatus `json:"token_info_retrieval,omitempty"`
+	MediaURLsRetrieval                    PipelineStepStatus `json:"media_urls_retrieval,omitempty"`
+	AnimationContentHeaderValueRetrieval  PipelineStepStatus `json:"animation_content_header_value_retrieval,omitempty"`
+	AnimationReaderRetrieval              PipelineStepStatus `json:"animation_reader_retrieval,omitempty"`
+	AnimationOpenseaFallback              PipelineStepStatus `json:"animation_opensea_fallback,omitempty"`
+	AnimationDetermineMediaTypeWithReader PipelineStepStatus `json:"animation_determine_media_type_with_reader,omitempty"`
+	AnimationAnimationGzip                PipelineStepStatus `json:"animation_animation_gzip,omitempty"`
+	AnimationStoreGCP                     PipelineStepStatus `json:"animation_store_gcp,omitempty"`
+	AnimationThumbnailGCP                 PipelineStepStatus `json:"animation_thumbnail_gcp,omitempty"`
+	AnimationLiveRenderGCP                PipelineStepStatus `json:"animation_live_render_gcp,omitempty"`
+	ImageContentHeaderValueRetrieval      PipelineStepStatus `json:"image_content_header_value_retrieval,omitempty"`
+	ImageReaderRetrieval                  PipelineStepStatus `json:"image_reader_retrieval,omitempty"`
+	ImageOpenseaFallback                  PipelineStepStatus `json:"image_opensea_fallback,omitempty"`
+	ImageDetermineMediaTypeWithReader     PipelineStepStatus `json:"image_determine_media_type_with_reader,omitempty"`
+	ImageAnimationGzip                    PipelineStepStatus `json:"image_animation_gzip,omitempty"`
+	ImageStoreGCP                         PipelineStepStatus `json:"image_store_gcp,omitempty"`
+	ImageThumbnailGCP                     PipelineStepStatus `json:"image_thumbnail_gcp,omitempty"`
+	ImageLiveRenderGCP                    PipelineStepStatus `json:"image_live_render_gcp,omitempty"`
+	NothingCachedWithErrors               PipelineStepStatus `json:"nothing_cached_errors,omitempty"`
+	NothingCachedWithoutErrors            PipelineStepStatus `json:"nothing_cached_no_errors,omitempty"`
+	CreateMediaFromCachedObjects          PipelineStepStatus `json:"create_media_from_cached_objects,omitempty"`
+	CreateRawMedia                        PipelineStepStatus `json:"create_raw_media,omitempty"`
+	SetUnknownMediaType                   PipelineStepStatus `json:"set_default_media_type,omitempty"`
+	MediaResultComparison                 PipelineStepStatus `json:"media_result_comparison,omitempty"`
+	UpdateTokenMetadataDB                 PipelineStepStatus `json:"update_token_metadata_db,omitempty"`
+	UpdateJobDB                           PipelineStepStatus `json:"update_job_db,omitempty"`
+}
+
+func (p PipelineMetadata) Value() (driver.Value, error) {
+	return json.Marshal(p)
+}
+
+func (p *PipelineMetadata) Scan(value interface{}) error {
+	if value == nil {
+		return nil
+	}
+	return json.Unmarshal(value.([]byte), p)
 }
 
 func TrackStepStatus(status *PipelineStepStatus) func() {
